@@ -1,6 +1,6 @@
 {
     "name": "@wxa/core",
-    "version": "2.0.7",
+    "version": "2.2.0",
     "description": "wxa core feature",
     "main": "./dist/main.js",
     "scripts": {
@@ -37,12 +37,12 @@
         "jest": "^23.6.0",
         "lint-staged": "^6.1.1",
         "uglifyjs-webpack-plugin": "^1.3.0",
-        "webpack": "^4.20.2",
+        "webpack": "^4.42.0",
         "webpack-cli": "^3.1.1"
     },
     "dependencies": {
         "deep-object-diff": "^1.1.0",
-        "lodash": "^4.17.11"
+        "lodash": "^4.17.15"
     },
     "lint-staged": {
         "*.js": [
@@ -58,5 +58,6 @@
     },
     "publishConfig": {
         "access": "public"
-    }
+    },
+    "gitHead": "b925c4fe80f493edd89c0c39a70e6eda3b97d915"
 }
