@@ -3,21 +3,62 @@ import logger from './helpers/logger';
 import fs from 'fs';
 import path from 'path';
 import shell from 'shelljs';
+import https from 'https';
+import {isUri} from 'valid-url';
+import inquirer from 'inquirer';
 
+let remoteMap = new Map([
+    ['github', 'https://github.com/wxajs'],
+    ['gitee', 'https://gitee.com/wxajs'],
+]);
+
+function getQA(templateConfigs) {
+    return [
+    {
+        type: 'input',
+        name: 'projectName',
+        message: '输入项目名',
+        validate: (input)=>{
+            return !(input == null || input === '');
+        },
+    },
+    {
+        type: 'list',
+        name: 'template',
+        message: '选择模板',
+        default: 'base',
+        choices: templateConfigs,
+    },
+    {
+        type: 'input',
+        name: 'appid',
+        message: '小程序APPID',
+        default: '',
+    },
+    ];
+}
 class Creator {
     constructor(cmd) {
         this.cmdOptions = cmd;
-        this.prefix = 'https://github.com/wxajs';
+        this.prefix = isUri(cmd.repo) ? cmd.repo : remoteMap.get(cmd.repo);
     }
 
-    run({projectName, template, ...rest}) {
+    async run() {
+        let configs = await this.getRemoteConfigs();
+        let options = await inquirer.prompt(getQA(JSON.parse(configs)));
+
+        this.create(options);
+    }
+
+    create({projectName, template, ...rest}) {
         this.clone(template, projectName, rest);
     }
 
     clone(template, name, rest) {
         let full = `${this.prefix}/wxa-templates.git`;
 
-        exec(`git clone ${full} ${name}`, function(err, stdout, stderr) {
+        logger.info('Downloading', `正在从 ${full} 下载模板`);
+        let childProcess = exec(`git clone ${full} ${name}`, function(err, stdout, stderr) {
             if (err) {
                 logger.error(err);
             } else {
@@ -54,6 +95,32 @@ class Creator {
                     logger.error(e);
                 }
             }
+        });
+
+        childProcess.on('error', (err) => {
+            logger.error(err);
+        });
+        childProcess.on('message', (err) => {
+            logger.info('Child Process', err);
+        });
+    }
+
+    getRemoteConfigs() {
+        let data = '';
+        return new Promise((resolve, reject) => {
+            let req = https.get(`${this.prefix}/wxa-templates/raw/master/configs.json`, (res)=>{
+                res.on('data', (d)=>{
+                    data += d.toString();
+                });
+                res.on('error', (e)=>{
+                    reject(e);
+                });
+                res.on('close', ()=>{
+                    resolve(data);
+                });
+            });
+
+            req.end();
         });
     }
 }
