@@ -1,10 +1,9 @@
 export default function(code, category='', path) {
     let temp = `
-        var wrapWxa = require('wxa://wxa_wrap.js').default;
-        wrapWxa(exports, "${category.toUpperCase()}", "${path}");
+var wrapWxa = require('wxa://wxa_wrap').default;
+wrapWxa(exports, "${category.toUpperCase()}", "${path}");
 
-        ${code}
-    `;
+${code}`;
 
     return temp;
 }
