@@ -3,6 +3,9 @@ import debugPKG from 'debug';
 import {AsyncSeriesHook} from 'tapable';
 import ProgressBar from '../helpers/progressTextBar';
 import SplitDeps from './splitDeps';
+import types from '../const/types';
+import {stuffEmptyAttributs} from './stuffEmptyAttributes';
+import transformPixelsToRpx from './pxtorpx';
 
 let debug = debugPKG('WXA:Optimizer');
 
@@ -33,6 +36,7 @@ export default class Optimizer {
 
         await Promise.all(optimizeTasks);
 
+
         if (!this.cmdOptions.watch) this.splitDeps.run(indexedMap);
     }
 
@@ -42,11 +46,27 @@ export default class Optimizer {
             debugger;
             return;
         }
+
         const text = path.relative(this.cwd, dep.src);
         this.progress.draw(text, 'Optimizing', !this.cmdOptions.verbose);
-        switch (this.wxaConfigs.target) {
-            case 'wxa':
-                this.doWxaOptimize(dep, indexedMap);
+
+        if (~types.WECHAT.concat(types.TT).indexOf(this.wxaConfigs.target)) {
+            this.doWxaOptimize(dep, indexedMap);
+        }
+
+        // stuff xml empty Attributes
+        if (
+            this.wxaConfigs.optimization.allowEmptyAttributes &&
+            ~['xml', 'wxml'].indexOf(dep.kind)
+        ) {
+            stuffEmptyAttributs(dep);
+        }
+
+        if (
+            ~['css'].indexOf(dep.kind) &&
+            this.wxaConfigs.optimization.transformPxToRpx
+        ) {
+            transformPixelsToRpx(dep, this.wxaConfigs);
         }
 
         await this.hooks.optimizeAssets.promise(dep);
