@@ -1,6 +1,11 @@
 import path from 'path';
 import logger from '../helpers/logger';
 
+const COLOR = {
+    SKIP: 1,
+    INIT: 0,
+};
+
 export default class SplitDeps {
     constructor({appConfigs, wxaConfigs, cwd, cmdOptions}) {
         this.cmdOptions = cmdOptions;
@@ -42,8 +47,22 @@ export default class SplitDeps {
         });
     }
 
+    clean(dep) {
+        dep.$depSplitColor = COLOR.INIT;
+        dep.childNodes.forEach((child)=> {
+            if (dep.$depSplitColor !== COLOR.INIT) {
+                this.clean(child);
+            }
+        });
+    }
+
     start(dep, pkg) {
+        dep.$depSplitColor = COLOR.SKIP;
+
         dep.childNodes.forEach((child, src)=>{
+            // circular dependency
+            if (child.$depSplitColor === COLOR.SKIP) return;
+
             if (
                 // an abstract file will never be output, but we still need to track it's childNodes.
                 child.isAbstract ||
@@ -53,24 +72,39 @@ export default class SplitDeps {
                 return this.start(child, pkg);
             }
 
-            // not match condition meanning all it's sub-child needn't handle.
-            // or the main packages depend on it.
-            // just stop the spliting loop.
-            if (
-                child.reference.size >= this.maxSplitDeps &&
-                this.getReferenceSize(child) > this.maxSplitDeps
-            ) return;
+            if (!this.ifMatchRule(child)) return;
 
-            if ( child.pret.isWXALib ) return;
-
-            if (this.isInMainPackage(child)) return;
             if (this.cmdOptions.verbose) logger.info('Find NPM need track to subpackages', child.src);
             // fulfill all condition just track all the sub-nodes without any hesitate.
             this.trackChildNodes(child, {output: dep.meta.outputPath, originOutput: dep.meta.outputPath, instance: dep, isSplitEntry: true}, pkg);
         });
     }
 
+    ifMatchRule(child) {
+        // not match condition meanning all it's sub-child needn't handle.
+        // or the main packages depend on it.
+        // just stop the spliting loop.
+        if (
+            child.reference.size >= this.maxSplitDeps &&
+            this.getReferenceSize(child) > this.maxSplitDeps
+        ) return false;
+
+        if (child.pret.isWXALib) return false;
+
+        if (this.isInMainPackage(child)) return false;
+
+        return true;
+    }
+
     isInMainPackage(child) {
+        if (!child.reference || !child.reference.size) {
+            // entry point.
+            return (
+                !this.subPages.some((sub)=>sub.reg.test(child.src)) ||
+                // if a node_module pkg is push in entry. then we always compiler its' to main-package.
+                this.NMReg.test(child.src)
+            );
+        }
         let refs = Array.from(child.reference);
         let inMain = [];
         refs.forEach(([src, mdl])=>{
@@ -84,10 +118,25 @@ export default class SplitDeps {
                 inMain.push(false);
             }
         });
-        return inMain.some((item)=>item);
+        let isInMainPkg = inMain.some((item)=>item);
+
+        if (!isInMainPkg && this.NMReg.test(child.src)) {
+            // deep check third party module.
+            for (let i = 0; i < refs.length; i ++) {
+                let parentNode = refs[i][1];
+                if (parentNode.isROOT) continue;
+
+                isInMainPkg = this.isInMainPackage(parentNode);
+                // if one of parent is in main, then just stop the loop.
+                if (isInMainPkg) break;
+            }
+        }
+
+        return isInMainPkg;
     }
 
     trackChildNodes(dep, parent, subpage) {
+        // if (/select\-revenue/g.test(dep.src)) debugger;
         // depth-first
         dep.$$isSplit = true;
         let {path: pkg} = subpage;
@@ -124,7 +173,27 @@ export default class SplitDeps {
         // update output path
         dep.output.add(newOutputPath);
 
-        if (dep.childNodes) dep.childNodes.forEach((child)=>this.trackChildNodes(child, {output: newOutputPath, originOutput: outputPath, instance: dep}, subpage));
+        if (dep.childNodes) {
+            dep.childNodes.forEach((child)=>{
+                // check node_modules's dependencies.
+                if (!this.ifMatchRule(child)) {
+                    // if (/wxa_wrap/.test(child.src)) {
+                    //     debugger;
+                    // }
+                    let originResolvedPath = './' + this.getResolved(outputPath, child.meta.outputPath);
+                    let newChildResolvedPath = './' + this.getResolved(newOutputPath, child.meta.outputPath);
+                    if (dep.kind === 'json') {
+                        originResolvedPath = this.getPathWithoutExtension(originResolvedPath);
+                        newChildResolvedPath = this.getPathWithoutExtension(newChildResolvedPath);
+                    }
+                    dep.code = dep.code.replace(new RegExp('["\']'+originResolvedPath+'["\']', 'gm'), '"'+newChildResolvedPath+'"');
+                    // debugger;
+                    return;
+                };
+
+                this.trackChildNodes(child, {output: newOutputPath, originOutput: outputPath, instance: dep}, subpage);
+            });
+        }
     }
 
     hasNoSplitReference(dep) {
