@@ -1,18 +1,19 @@
 import traverse from '@babel/traverse';
 import * as t from '@babel/types';
 import template from '@babel/template';
-import generate from '@babel/generator';
 import path from 'path';
 import DependencyResolver from '../../helpers/dependencyResolver';
 import debugPKG from 'debug';
 import logger from '../../helpers/logger';
+import {generateCodeFromAST} from '../../compilers/script';
 
 let debug = debugPKG('WXA:ASTManager');
 
 export default class ASTManager {
-    constructor(resolve, meta) {
+    constructor(resolve, meta, wxaConfigs) {
         this.resolve = resolve;
         this.meta = meta;
+        this.wxaConfigs = wxaConfigs;
 
         this.modulesPath = path.join(meta.current, 'node_modules');
         this.npmPath = path.join(meta.output.path, 'npm');
@@ -72,12 +73,19 @@ export default class ASTManager {
 
         let self = this;
         let libs = [];
+        let importStatement = [
+            t.callExpression.name,
+            t.importDeclaration.name,
+            t.exportAllDeclaration.name,
+            t.exportNamedDeclaration.name].join('|');
+
         traverse(mdl.ast, {
-            'CallExpression|ImportDeclaration': (path) => {
+            [importStatement]: (path) => {
                 let dep;
                 let typeOfPath;
                 const StringLiteralRequire = 'StringLiteralRequire';
                 const ImportDeclaration = 'ImportDeclaration';
+                const ExportDeclaration = 'ExportDeclaration';
 
                 // commandJS module
                 if (
@@ -102,6 +110,13 @@ export default class ASTManager {
                 ) {
                     dep = path.node.source.value;
                     typeOfPath = ImportDeclaration;
+                } else if (
+                    (t.isExportAllDeclaration(path.node) || t.isExportNamedDeclaration(path.node)) &&
+                    t.isLiteral(path.node.source) &&
+                    path.node.source.value
+                ) {
+                    dep = path.node.source.value;
+                    typeOfPath = ExportDeclaration;
                 } else {
                     return;
                 }
@@ -126,20 +141,40 @@ export default class ASTManager {
                         case StringLiteralRequire:
                             // path.replaceWithSourceString(`require("${resolved}")`);
                             path.replaceWith( template(`require(SOURCE)`)({SOURCE: t.stringLiteral(resolved)}) );
+                            path.stop();
                             break;
                         case ImportDeclaration:
                             path.get('source').replaceWith(t.stringLiteral(resolved));
+                            path.skip();
+                            break;
+                        case ExportDeclaration:
+                            path.get('source').replaceWith(t.stringLiteral(resolved));
+                            path.skip();
                             break;
                     }
-                    path.stop();
                 } catch (e) {
                     logger.error('解析失败', e);
                     debug('resolve fail %O', e);
                 }
             },
-            'IfStatement': (path) => {
+            [t.ifStatement.name]: (path) => {
                 // check Unreachable code
                 this.checkUnreachableCode(path);
+            },
+            [t.assignmentExpression.name]: (path) => {
+                // inject platform env to runtime app.
+                if (
+                    t.isMemberExpression(path.node.left) &&
+                    t.isThisExpression(path.node.left.object) &&
+                    t.isIdentifier(path.node.left.property, {name: '__WXA_PLATFORM__'})
+                ) {
+                    let rightExpressionPath = path.get('right');
+
+                    rightExpressionPath.replaceWith(
+                        t.stringLiteral(this.wxaConfigs.target)
+                    );
+                    // debugger;
+                }
             },
         });
 
@@ -149,7 +184,9 @@ export default class ASTManager {
 
         libs = libs.concat(wxaSourceLibs);
         // generate module code.
-        mdl.code = this.generate(mdl).code;
+        let {code, map} = this.generate(mdl);
+        mdl.code = code;
+        // mdl.sourceMap = map;
         delete mdl.ast;
         return libs;
     }
@@ -196,8 +233,6 @@ export default class ASTManager {
     generate(mdl) {
         debug('generate start');
         if (mdl.ast == null) return;
-
-        // debug('module to generate %O', mdl.ast);
-        return generate(mdl.ast, {});
+        return generateCodeFromAST(mdl.ast, {}, mdl);
     }
 }
