@@ -109,7 +109,7 @@ export function amazingCache(params, needCache) {
     let defaultOpts = {
         directory: true,
         identifier: JSON.stringify({
-            '@webank/wxa-cli': pkg.version,
+            '@wxa/cli2': pkg.version,
             'env': process.env.NODE_ENV || 'development',
         }),
     };
@@ -154,4 +154,22 @@ export function promiseSerial(funs) {
     return funs.reduce((promise, fun)=>{
         return promise.then((result)=>fun().then(Array.prototype.concat.bind(result)));
     }, Promise.resolve([]));
+}
+
+export function getClassSet(classStr) {
+    let classList = [];
+    if (classStr && typeof classStr === 'string') {
+        classList = classStr.split(' ');
+    }
+    return new Set(classList);
+}
+export function addClass(classStr, newClass) {
+    let classSet = getClassSet(classStr);
+    classSet.add(newClass);
+    return Array.from(classSet);
+}
+export function removeClass(classStr, destClass) {
+    let classSet = getClassSet(classStr);
+    classSet.delete(destClass);
+    return Array.from(classSet);
 }
