@@ -19,10 +19,24 @@ class SassCompiler {
         if(this.configs == null) this.configs = configs || {};
     }
 
-    async parse(mdl, cmdConfigs, compilation) {
+    async parse(mdl, cmdOptions, compilation) {
         debug('sass module parse started %O', mdl);
+        // custom outputPath
+        if(mdl.meta) {
+            let source = path.parse(mdl.meta.source);
+            mdl.meta.source = source.dir + path.sep + source.name + '.css';
+
+            let output = path.parse(mdl.meta.outputPath);
+            mdl.meta.outputPath = output.dir + path.sep + output.name + '.css'
+        }
 
         let configs  = this.configs;
+
+        if (cmdOptions && cmdOptions.sourceMap) {
+            // inline css sourcemap.
+            configs.sourceMap = cmdOptions.sourceMap;
+            configs.sourceMapEmbed = true;
+        }
 
         let ret = await this.render(mdl.content || null, mdl.meta.source, configs);
         
@@ -32,15 +46,6 @@ class SassCompiler {
             ret.stats.includedFiles.forEach((file)=>{
                 mdl.dependency(file, compilation);
             });
-        }
-
-        // custom outputPath
-        if(mdl.meta) {
-            let source = path.parse(mdl.meta.source);
-            mdl.meta.source = source.dir + path.sep + source.name + '.css';
-
-            let output = path.parse(mdl.meta.outputPath);
-            mdl.meta.outputPath = output.dir + path.sep + output.name + '.css'
         }
 
         return {ret, code: mdl.code};
