@@ -1,9 +1,10 @@
-import {classFactory, methodDescriptorGenerator} from './helpers';
+import {classFactory} from './helpers';
 import promisify from '../utils/promisify';
 import {eventbus} from '../utils/eventbus';
 import {router} from '../utils/router';
 import wxapi from '../utils/wxapi';
 import {storage} from '../utils/storage';
+import {sessionStorage} from '../utils/sessionStorage';
 import debounce from 'lodash/debounce';
 import throttle from 'lodash/throttle';
 import * as helpers from '../utils/helpers';
@@ -15,6 +16,7 @@ const Eventbus = classFactory('$eventbus', eventbus);
 const Router = classFactory('$router', router);
 const Wxapi = classFactory('$wxapi', wxapi(wx));
 const Storage = classFactory('$storage', storage);
+const SessionStorage = classFactory('$sessionStorage', sessionStorage);
 const Utils = classFactory('$utils', {
     debounce,
     promisify,
@@ -25,38 +27,50 @@ const Fetch = classFactory('$fetch', fetch);
 const Mixins = (...args)=>classFactory('mixins', [mixin({mixins: args})]);
 
 // Page and App level class Decorators.
-const Page = (classDescriptor)=>{
+const Page = (classDescriptor) => {
     classDescriptor = Utils(classDescriptor);
     classDescriptor = Storage(classDescriptor);
+    classDescriptor = SessionStorage(classDescriptor);
     classDescriptor = Wxapi(classDescriptor);
     classDescriptor = Router(classDescriptor);
     classDescriptor = Eventbus(classDescriptor);
     classDescriptor = Fetch(classDescriptor);
 
-    let {elements} = classDescriptor;
+    // let {elements} = classDescriptor;
 
-    // 兼容wxa1.0 还是挂载一个app对象到每个页面实例
-    return {
-        ...classDescriptor,
-        elements: elements.concat([methodDescriptorGenerator('$app', getApp())]),
-    };
+    return classDescriptor;
 };
 
-const App = (classDescriptor)=> {
+const Component = (classDescriptor) => {
     classDescriptor = Utils(classDescriptor);
     classDescriptor = Storage(classDescriptor);
+    classDescriptor = SessionStorage(classDescriptor);
     classDescriptor = Wxapi(classDescriptor);
     classDescriptor = Router(classDescriptor);
     classDescriptor = Eventbus(classDescriptor);
     classDescriptor = Fetch(classDescriptor);
-    // console.log(classDescriptor);
+
+    return classDescriptor;
+};
+
+const App = (classDescriptor) => {
+    classDescriptor = Utils(classDescriptor);
+    classDescriptor = Storage(classDescriptor);
+    classDescriptor = SessionStorage(classDescriptor);
+    classDescriptor = Wxapi(classDescriptor);
+    classDescriptor = Router(classDescriptor);
+    classDescriptor = Eventbus(classDescriptor);
+    classDescriptor = Fetch(classDescriptor);
+
     return classDescriptor;
 };
 
 export {
     Page,
     App,
+    Component,
     Storage,
+    SessionStorage,
     Wxapi,
     Router,
     Eventbus,
